% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-effect.R
\name{between_effect}
\alias{between_effect}
\title{Between-subjects effects}
\usage{
between_effect(
  model,
  type = c("all", "main", "simple"),
  method = c("revpairwise", "pairwise", "dunnett", "trt.vs.ctrl"),
  adjust = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr", "none"),
  ref = 1,
  se = TRUE,
  language = c("en", "chn"),
  digits.effect = 2,
  digits.pvalue = 3,
  table.number = NULL
)
}
\arguments{
\item{model}{a object from \link{repeated_measures_long} or \link{repeated_measures_wide}.}

\item{type}{a character specifying the effect type, 'main' for main effects,
'simple' for simple effects. 'all' for main and simple effects.}

\item{method}{a character value giving the root name of a contrast method.}

\item{adjust}{a character, adjust P-values for multiple comparisons, see \link{p.adjust}.}

\item{ref}{integer(s) or character(s) specifying which level(s) to use as the
reference. Character values must exactly match elements of levs.}

\item{se}{a logical variable specifying whether to shou standard errors or confidence intervals.}

\item{language}{language, 'en' for English, 'chn' for Chinese.}

\item{digits.effect}{digits for effect, default 2.}

\item{digits.pvalue}{digits for p value, default 3.}

\item{table.number}{table number.}
}
\value{
a data.frame with class 'rmanova'.
}
\description{
Between-subjects effects
}
\examples{
model <- repeated_measures_long(data = ldata,
                                between = treatment,
                                within = time,
                                response = score,
                                subject = subject)

# Main effects and simple effects
between_effect(model, type = "all")

# Main effects
between_effect(model, type = "main")

# Simple effects
between_effect(model, type = "simple")

# Write to docx.
# results <- between_effect(model, type = "all", table.number = 1)
# write_docx(results, path = "Between effect.docx")
}
