% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{body_add_par2}
\alias{body_add_par2}
\title{Add paragraphs of text in a 'Word' document}
\usage{
body_add_par2(x, value, style = NULL, pos = "after")
}
\arguments{
\item{x}{a docx device.}

\item{value}{a character}

\item{style}{paragraph style name}

\item{pos}{where to add the new element relative to the cursor, one of "after", "before", "on"}
}
\description{
Add paragraphs of text in a 'Word' document
}
\keyword{internal}
