% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_plot}
\alias{gg_plot}
\title{Plot for repeated measures ANOVA}
\usage{
gg_plot(
  model,
  error.type = c("se", "ci", "sd", "none"),
  method = "revpairwise",
  adjust = "bonferroni",
  ref = 1,
  signif = TRUE,
  error.width = 0.1,
  line.size = 0.25,
  point.size = 1.5,
  vjust.signif = -0.5,
  family = "serif",
  legend.position = NULL,
  ...
)
}
\arguments{
\item{model}{a object from \link{repeated_measures_long} or \link{repeated_measures_wide}}

\item{error.type}{the type of error bar, one of 'se', 'ci', 'sd', and 'none'.}

\item{method}{a character value giving the root name of a contrast method, see \link{contrast} or \link{contrast-methods}.}

\item{adjust}{a character, adjust P-values for multiple comparisons, see \link{contrast} or \link{p.adjust}.}

\item{ref}{integer(s) or character(s) specifying which level(s) to use as the reference. Character values must exactly match elements of levs}

\item{signif}{a logical variable indicating whether to show significance codes.}

\item{error.width}{width of error bar.}

\item{line.size}{line size.}

\item{point.size}{point size.}

\item{vjust.signif}{vjust for significance codes.}

\item{family}{font family.}

\item{legend.position}{legend's position.}

\item{...}{unused.}
}
\value{
a object of ggplot2.
}
\description{
Plot for repeated measures ANOVA
}
\examples{
model <- repeated_measures_long(data = ldata,
                                between = treatment,
                                within = time,
                                response = score,
                                subject = subject)

gg_plot(model, error.type = "se")
gg_plot(model, error.type = "ci")
gg_plot(model, error.type = "none")
}
