% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ldata}
\alias{ldata}
\title{Long format data}
\format{
A data frame with 180 rows and 4 variables:
\describe{
  \item{subject}{Subject.}
  \item{treatment}{Treatment.}
  \item{time}{Time at baseline, 3 months, and 6 months.}
  \item{score}{Score.}
}
}
\usage{
ldata
}
\description{
Long format data
}
\keyword{datasets}
