% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.afex_aov}
\alias{print.afex_aov}
\title{Print  object}
\usage{
\method{print}{afex_aov}(x, ...)
}
\arguments{
\item{x}{a object of 'afex_aov".}

\item{...}{further arguments.}
}
\description{
Print  object
}
\keyword{internal}
