% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rmanova}
\alias{print.rmanova}
\title{Print 'rmanova' object}
\usage{
\method{print}{rmanova}(x, ...)
}
\arguments{
\item{x}{a object of 'rmanova".}

\item{...}{further arguments.}
}
\description{
Print 'rmanova' object
}
\keyword{internal}
