% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.srrma}
\alias{print.srrma}
\title{Print  object}
\usage{
\method{print}{srrma}(x, ...)
}
\arguments{
\item{x}{a object of 'srrma".}

\item{...}{further arguments.}
}
\description{
Print  object
}
\keyword{internal}
