% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{repeated_measures_wide}
\alias{repeated_measures_wide}
\title{Repeated measures ANOVA for wide format data}
\usage{
repeated_measures_wide(
  data,
  between = NULL,
  responses = NULL,
  covariates = NULL,
  subject = NULL,
  interaction = TRUE
)
}
\arguments{
\item{data}{a data.frame is in the so-called wide format.}

\item{between}{between-subject variables.}

\item{responses}{dependent variable.}

\item{covariates}{covariates.}

\item{subject}{a subject is an observational unit that can be considered independent of other subjects.}

\item{interaction}{a logical variable specifying whether to include interaction terms.}
}
\value{
a object of repeated measures ANOVA.
}
\description{
Repeated measures ANOVA for wide format data
}
\examples{
repeated_measures_wide(data = wdata,
                       subject = subject,
                       between = treatment,
                       responses = T0:T6)
}
