% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmanova-effect.R
\name{rmanova_effect}
\alias{rmanova_effect}
\title{Effects of Repeated-measures ANOVA}
\usage{
rmanova_effect(
  model,
  digits.effect = 2,
  digits.pvalue = 3,
  language = "en",
  table.number = NULL
)
}
\arguments{
\item{model}{a object from \link{rmanova_long} or \link{rmanova_wide}.}

\item{digits.effect}{digits for effect, default 2.}

\item{digits.pvalue}{digits for p value, default 3.}

\item{language}{language, 'en' for English, 'chn' for Chinese.}

\item{table.number}{table number.}
}
\value{
a data.frame.
}
\description{
Effects of Repeated-measures ANOVA
}
\examples{
model <- rmanova_long(data     = ldata,
                      between  = treatment,
                      within   = time,
                      response = score,
                      subject  = subject)

rmanova_effect(model)
}
