% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rmanova_long}
\alias{rmanova_long}
\title{Repeated measures ANOVA for long format data}
\usage{
rmanova_long(
  data,
  between = NULL,
  within = NULL,
  response = NULL,
  covariates = NULL,
  subject = NULL,
  interaction = TRUE
)
}
\arguments{
\item{data}{a data.frame is in the so-called long format.}

\item{between}{between-subject variables.}

\item{within}{within-subjects variables.}

\item{response}{dependent variable.}

\item{covariates}{covariates.}

\item{subject}{a subject is an observational unit that can be considered independent of other subjects.}

\item{interaction}{a logical variable specifying whether to include interaction terms.}
}
\value{
a object of repeated measures ANOVA.
}
\description{
Repeated measures ANOVA for long format data
}
\examples{
rmanova_long(data     = ldata,
             between  = treatment,
             within   = time,
             response = score,
             subject  = subject)

rmanova_long(data     = ldata,
             between  = "treatment",
             within   = "time",
             response = "score",
             subject  = "subject")
}
