% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdata}
\alias{wdata}
\title{Wide format data}
\format{
A data frame with 60 rows and 5 variables:
\describe{
  \item{subject}{Subject.}
  \item{treatment}{Treatment.}
  \item{T0}{Score at baseline.}
  \item{T3}{Score at 3th months.}
  \item{T6}{Score at 6th months.}
}
}
\usage{
wdata
}
\description{
Wide format data
}
\keyword{datasets}
